studygenie/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.4
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
googletrans==4.0.0rc1
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2